Move title component
